Changes:
- split base.py
- don't require settings being a pydantic settings object
- add get_value_from_settings helper for handling attribute based
  settings as well dictionaries
- bump version
