+++ monkay/__init__.py
@@ -3,12 +3,19 @@
 # SPDX-License-Identifier: BSD-3-Clauses
 
 from .base import (
+    InGlobalsDict,
+    absolutify_import,
+    get_value_from_settings,
+    load,
+    load_any,
+)
+from .core import (
+    Monkay,
+)
+from .types import (
     PRE_ADD_LAZY_IMPORT_HOOK,
     DeprecatedImport,
     ExtensionProtocol,
-    Monkay,
-    load,
-    load_any,
 )
 
 __all__ = [
@@ -20,4 +27,6 @@ __all__ = [
     "load",
     "load_any",
     "absolutify_import",
+    "InGlobalsDict",
+    "get_value_from_settings",
 ]
