+++ pyproject.toml
@@ -14,22 +14,23 @@ authors = [
   { name = "Tiago Silva", email = "tiago.arasilva@gmail.com" }
 ]
 classifiers = [
-  "Development Status :: 4 - Beta",
+  "Operating System :: OS Independent",
+  "Topic :: Software Development :: Libraries :: Python Modules",
+  "Topic :: Software Development :: Libraries",
+  "Topic :: Software Development",
+  "Typing :: Typed",
+  "Development Status :: 5 - Production/Stable",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
   "Programming Language :: Python :: Implementation :: CPython",
   "Programming Language :: Python :: Implementation :: PyPy",
   "License :: OSI Approved :: BSD License"
 ]
 dependencies = []
-
-[project.optional-dependencies]
-settings = [
-  "pydantic-settings"
-]
 
 [project.urls]
 Documentation = "https://github.com/devkral/monkay#readme"
